requirements.txt
# 📁 Ruta: conexapi_backend/requirements.txt  
# 🧾 Archivo: requirements.txt  
# 🎯 Objetivo: Listar las dependencias necesarias del backend para instalación en entorno local y despliegue.  
# 📌 Estado: Validado y compatible con FastAPI + MariaDB (usando aiomysql).

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
mysql-connector-python==9.0.0 #para conectar a MySQL
pymysql
sqlalchemy==2.0.25
alembic==1.13.1
aiomysql==0.2.0          # Cliente asincrónico, compatible con SQLAlchemy 2 y async
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Validation & Serialization
email-validator==2.1.0.post1
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# Logging
structlog==23.2.0

# Environment
python-dotenv==1.0.0

# ASGI Server Production
gunicorn==21.2.0

# Monitoring (optional)
prometheus-client==0.19.0

# GitHub Actions Dependencies
coverage==7.3.2